requirements.txt
Django==3.0.6
djangorestframework==3.11.0
django-rest-auth==0.9.5
django-allauth==0.41.0
django-extensions==2.2.9
django-sslserver==0.22
